+++ CHANGELOG.md
@@ -2,9 +2,10 @@
 
 ### Features
 
- - Add ability to set each Graph Widget series type to either Line or Column
+ - Add Function: Regulate pH and Electrical Conductivity
  - Add Input: MCP3008 (adafruit-circuitpython-mcp3xxx library)
  - Add Input: On/Off Output State
+ - Add ability to set each Graph Widget series type to either Line or Column
  - Add Temperature Calibration for Atlas Scientific PT-1000 Input
 
 ### Bugfixes